+++ src/main/java/com/cisowski/schoolmanagement/schedule/service/ScheduleService.java
@@ -11,4 +11,5 @@ public interface ScheduleService {
     ScheduleDetailedResponse getSchedule(Integer scheduleId);
     List<ScheduleSummaryResponse> getScheduleByDayOfWeek(Integer scheduleVersionId, Integer dayOfWeek, boolean needsFiltering, Integer userId);
     List<ScheduleEntity> fetchSchedulesByClassroomId(Integer classroomId);
+    void cancelSchedule(Integer scheduleId, String reason);
 }